+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Development version
+
+Bug fixes:
+
+- Ensure that GeoDataFrame created from DataFrame is a copy, not a view (#2667)
+
 ## Version 0.12.2 (December 10, 2022)
 
 Bug fixes: